Add redis-check middleware with an error template
Add new middleware to render an error when the connection to Redis is lost.
To enable re-use of the current template the locals.js needed to be split up,
so there are 'pre session' and 'post-session' locals
