+++ src/middleware/locals.js
@@ -1,10 +1,8 @@
-const { get, map } = require('lodash')
+const { map } = require('lodash')
 
-const GLOBAL_NAV_ITEMS = require('../apps/global-nav-items')
 const logger = require('../config/logger')
 const config = require('../config')
 const { version } = require('../../package.json')
-const { filterNonPermittedItem } = require('../modules/permissions/filters')
 
 let webpackManifest = {}
 
@@ -14,28 +12,10 @@ try {
   logger.error('Manifest file is not found. Ensure assets are built.')
 }
 
-function getActiveHeaderKey (requestPath, permittedNavItems) {
-  if (requestPath.startsWith('/support')) {
-    return 'datahub-support'
-  } else if (requestPath === '/profile') {
-    return 'datahub-profile'
-  } else {
-    for (const { path, headerKey } of permittedNavItems) {
-      if (requestPath.startsWith(path)) {
-        return headerKey
-      }
-    }
-  }
-}
-
 module.exports = function locals (req, res, next) {
   const baseUrl = `${(req.encrypted ? 'https' : req.protocol)}://${req.get('host')}`
-  const userPermissions = get(res, 'locals.user.permissions')
-  const userProfile = config.oauth.bypassSSO ? null : get(req.session, 'userProfile')
-  const permittedApplications = get(userProfile, 'permitted_applications', [])
-  const permittedNavItems = GLOBAL_NAV_ITEMS.filter(filterNonPermittedItem(userPermissions))
 
-  res.locals = Object.assign({}, res.locals, {
+  Object.assign(res.locals, {
     APP_VERSION: version,
     BASE_URL: baseUrl,
     CANONICAL_URL: baseUrl + req.path,
@@ -45,16 +25,6 @@ module.exports = function locals (req, res, next) {
     GOOGLE_TAG_MANAGER_KEY: config.googleTagManagerKey,
     IS_XHR: req.xhr,
     QUERY: req.query,
-    PERMITTED_APPLICATIONS: (config.oauth.bypassSSO ? [{ key: 'datahub-crm' }] : permittedApplications),
-    ALLOWED_APPS: permittedNavItems.reduce((apps, { headerKey }) => {
-      headerKey && apps.push(headerKey)
-      return apps
-    }, []),
-    ACTIVE_KEY: getActiveHeaderKey(req.path, permittedNavItems),
-
-    getMessages () {
-      return req.flash()
-    },
 
     getPageTitle () {
       const items = res.breadcrumb().map(item => item.text)
