+++ src/server.js
@@ -19,6 +19,7 @@ const currentJourney = require('./modules/form/current-journey')
 const nunjucks = require('./config/nunjucks')
 const headers = require('./middleware/headers')
 const locals = require('./middleware/locals')
+const userLocals = require('./middleware/user-locals')
 const metadata = require('./lib/metadata')
 const user = require('./middleware/user')
 const auth = require('./middleware/auth')
@@ -30,6 +31,7 @@ const ssoBypass = require('./middleware/sso-bypass')
 const logger = require('./config/logger')
 const basicAuth = require('./middleware/basic-auth')
 const features = require('./middleware/features')
+const redisCheck = require('./middleware/redis-check')
 const reporter = require('./lib/reporter')
 
 const routers = require('./apps/routers')
@@ -53,11 +55,6 @@ if (!config.isDev) {
 }
 
 app.use(cookieParser())
-app.use(sessionStore)
-
-app.use(bodyParser.urlencoded({ extended: true, limit: '1mb' }))
-app.use(bodyParser.json())
-
 app.use(compression())
 
 app.set('view engine', 'njk')
@@ -89,6 +86,13 @@ app.use('/assets', express.static(path.join(config.root, 'node_modules/govuk-fro
 app.use(title())
 app.use(breadcrumbs.init())
 app.use(breadcrumbs.setHome())
+app.use(locals)
+app.use(redisCheck)
+
+app.use(sessionStore)
+
+app.use(bodyParser.urlencoded({ extended: true, limit: '1mb' }))
+app.use(bodyParser.json())
 
 app.use(currentJourney())
 
@@ -99,7 +103,7 @@ app.use(basicAuth)
 app.use(auth)
 app.use(user)
 app.use(features)
-app.use(locals)
+app.use(userLocals)
 app.use(headers)
 app.use(store())
 app.use(csrf())
