+++ src/templates/_macros/common/local-header.njk
@@ -14,7 +14,7 @@
  #}
 {% macro LocalHeader(props) %}
   {% set breadcrumbs = props.breadcrumbs | default(getBreadcrumbs()) %}
-  {% set messages = props.messages | default(getMessages()) %}
+  {% set messages = props.messages | default(getMessages() if getMessages) %}
   {% set modifier = props.modifier | concat('') | reverse | join(' c-local-header--') if props.modifier %}
 
   {% if props %}
